fixing typo in torch backend subtract() | numpy-array-api test should pass
Former-commit-id: d6d27963d03394f6fcae0b3e51fab4c39895b0c4

+++ ivy/functional/backends/torch/elementwise.py
@@ -421,7 +421,7 @@ def subtract(
         return torch.subtract(x1, x2, out=out)
     return torch.subtract(
         x1 if isinstance(x1, torch.Tensor) else torch.tensor(x1, dtype=x2.dtype),
-        x2 if isinstance(x2, torch.Tensor) else torch.tensor(x2, dtype=x2.dtype),
+        x2 if isinstance(x2, torch.Tensor) else torch.tensor(x2, dtype=x1.dtype),
     )
